#198: This changeset allows centering of both the text and watermark primitives (and will work for any others I choose to add in future). I will be updating the UI for this in the near future to combine both the absolute and relative positioning settings.

+++ Source/Application/Controllers/VTLightSensorController.h
@@ -19,7 +19,7 @@
   BOOL          enabled;
   BOOL          canEnable;
   
-  int           sensitivity;
+  float           sensitivity;
   int           stable;
   int           left_sum, right_sum;
   int           count;
@@ -38,8 +38,8 @@
   // Getters and setters
 - (BOOL) isEnabled;
 - (void) setIsEnabled: (BOOL) enableValue;
-- (int) sensorSensitivity;
-- (void) setSensorSensitivity: (int) sensitivityValue;
+- (float) sensorSensitivity;
+- (void) setSensorSensitivity: (float) sensitivityValue;
 
 
 @end
